Add GET controller for CRUD
This commit also adds a new "OkContent" response that can handle a
content. This response will be ignore on serialization and only the
content will be the result of the serialization thanks to the
OkContentNormalizer.

For obvious customization need, it also comes with a new feature:
serialization context, which only support groups for the moment but is
extensible.