Add a complete feature of filters to Melodiia
Filters were just little class without possibilities. This commit
adds a complete system for filters based on form.

This means that it can:
- Validate filters
- Answer consistent API responses

Filters are now super easy to create, this commit adds a documentation
about.

Please consider this commit highly breaks the compatibility. Some of
this breaks could have been avoided but it is done this way because
Melodiia is not stable ATM.

Note about filters that only support Doctrine: for now there's nothing
else than Doctrine supported by Melodiia. It could be nice to add a way
that allow us to bring more inside Melodiia. But it's not a nice DX
considering it needs us to drop some type hinting and we have no
replacement for now.