Add new option custom data mapper on apitype
The reason of this addition is that the ApiType is a great shortcut. But
defining the data mapper was a pain considering it's a custom one also
used for instanciation of items.

This new options makes customization smooth.

It also adds a new interface because we need to typehint the
customDataMapper and the method createObject is definitely important.