Improve input management
Problem: depending on the JSON input, sometimes the CRUD controllers
were returning 500 error while the actual correct code is 400.

Solution: catch the decoding exception.

This commit also improvove the management of the form creation by using
a new controller. This controller defines a protected method, if you
extends a melodiia controller you will be able to change the way it
behave on creation/update.