Make Melodiia compatible Symfony 5.0
This commit add test to lower version of dependencies, so it also adds
minimum low vesions of those.

doctrine/annotation is a special case because it's a dependency of
dependency. Also swagger php will be removed of Melodiia in next
release, so I do not add it to normal requirements.