Add new concept of id resolver
Problem:
Our controller had hardcoded id getter in the request, but it's actually
quite normal to have the ability of changing this id name inside the
route.

Solution:
After thinking first to make controller not final with a protected
method inside, I finally did this implementation of an id resolver. It
could even be configurable in the future with a new configurable class
id resolver.

This is a fix for #64