feat(crud): security is optionnal
This commit removes a mandatory dependency for CRUD controllers. An
exception will be throw in case the user didn't configure the security
component but security for Melodiia.

It comes with a huge BC break because it required to change the order of
the properties while those controllers are designed to be extendable. I
allowed this break to happen because we break on many things for the
rebranding.