refactor(doc): remove documentation management
Fix #5

This commit mostly removes features from Melodiia. I do it because after
experimenting automatic documentation I stated that this is mostly
stupid.

Automatic documentation generates inaccurate documentation. (we had
tests that the doc was at least valid and were putting effort so it is
valid) The point is that it was kinda useless and so incomplete. It was
an absolute requirement to use the phpdoc to make it right.

The problem is that writting phpdoc for the swagger documentation is an
infinite pain. Writting it in yaml is so much easier.

So this commit removes automatic doc and support for phpdocs in favor of
adding a new controller that will automatically load a given yaml file.