fix(filters): GET request support in ApiRequestHanler
After a some fixes and refactoring the filter system were not working anymore. This commit fixes the problem.

Please notice that it's not possible to send content in GET. This is
invalid on REST POV but could be required in some extremely rare
situation. I suggest to redefine the RequestApiHandler in those cases.