feat: remove useless override of Symfony
This override was originally made to make melodiia easier to use, but
it's finally just a stupid magic overhead that break the DX because in
some cases you need to add configuration to disable things.

It was going directly against what we want to do: to let the user be the
master of its code.