feat: new way to configure doc path
In the case of full managed applications, the app is often build in a
folder then moved to another. The routing is resolved and cached at
build time.

Therefore using the routing is not a good option to give a static path.
To avoid that we use the configuration instead.